requirements.txt
dateparser==0.5.0
Django==1.8.2
django-bootstrap3==7.1.0
django-google-maps==0.6.0
django-smtp-ssl==1.0
jdatetime==1.8.1
MySQL-python==1.2.5
Pillow==3.3.0
python-dateutil==2.6.0
pytz==2016.7
regex==2016.11.21
reportlab==3.3.0
ruamel.ordereddict==0.4.9
ruamel.yaml==0.13.2
six==1.10.0
South==1.0.2
typing==3.5.2.2
tzlocal==1.3
umalqurra==0.2
Xlsxcursor==0.1.7
XlsxWriter==0.9.3
